Added rapid modules for choreographed movement/ROS.
Choreo raises and lowers the arm. It was created for the Ford demo
videos.
Main module is just a switch to run choreo or ros_relay.
Ros_relay is currently just a tcp hello world.
State_server is the beginning of the server for sending robot state
data to ROS industrial.