Updated fixed fram no-arm CAD files and simmechanics files.
Broke out the wheels and casters in the base model instead of
using a sub-assembly because the simmechanics file generated
would create a messed-up URDF in which there was one caster with
two links to it from the base.

Base now has caster and caster wheel as parts in it instead of a
sub-assy.