Abby's URDF is looking pretty good.
This URDF looks good in the graph visualizer (no arm). I need to
change the joint defs for the casters and wheels so they can turn
and will probably need to make some tweaks when I load this in rviz
and gazebo. Then I'll need to add the arm w/ the xacro macro.